% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_aux.R
\name{apply_common_theme_and_labs}
\alias{apply_common_theme_and_labs}
\title{Función auxiliar para aplicar temas y etiquetas comunes}
\usage{
apply_common_theme_and_labs(
  p,
  settings,
  plot_type,
  add_grid_lines,
  show_background,
  title,
  subtitle,
  caption,
  x_label,
  y_label,
  base_theme_fun,
  grid_linetype,
  grid_linewidth,
  axis_line_linewidth,
  panel_background_map_specific = FALSE,
  text_size = 12
)
}
\arguments{
\item{p}{Objeto ggplot2 actual.}

\item{settings}{Lista de configuraciones de estilo (colores, fuentes, etc.).}

\item{plot_type}{Tipo de gráfico (para ajustes condicionales de ejes/cuadrícula).}

\item{add_grid_lines}{Lógico, si mostrar líneas de cuadrícula.}

\item{show_background}{Lógico, si mostrar el fondo del panel.}

\item{title}{Título del gráfico.}

\item{subtitle}{Subtítulo del gráfico.}

\item{caption}{Leyenda del gráfico.}

\item{x_label}{Nombre de la variable del eje x (para etiquetas).}

\item{y_label}{Nombre de la variable del eje y (para etiquetas).}

\item{base_theme_fun}{Función de tema base de ggplot2 (ej. `ggplot2::theme_void`).}

\item{grid_linetype}{Tipo de línea para la cuadrícula principal.}

\item{grid_linewidth}{Ancho de línea para la cuadrícula principal.}

\item{axis_line_linewidth}{Ancho de línea para los ejes.}

\item{panel_background_map_specific}{Lógico, si el tema del panel tiene lógica especial para mapas.}
}
\value{
Objeto ggplot2 con el tema y las etiquetas aplicadas.
}
\description{
Encapsula la lógica de ggplot2::theme() y ggplot2::labs() para reducir la duplicación de código.
}
