% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_aux.R
\name{get_artist_settings}
\alias{get_artist_settings}
\title{Configura los estilos visuales según el artista y obra}
\usage{
get_artist_settings(
  artist = c("da_vinci", "michelangelo", "rembrandt", "van_gogh", "monet", "banksy",
    "dali", "gentileschi", "miro", "kandinsky", "warhol"),
  work
)
}
\arguments{
\item{artist}{Nombre del artista: "da_vinci" o "michelangelo".}

\item{work}{Nombre de la obra específica del artista.}
}
\value{
Una lista con los parámetros del estilo visual.
}
\description{
Devuelve una lista con parámetros visuales como paleta, tipografía, colores y más, según el artista y la obra elegida.
}
