% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_styles.R
\name{load_all_artist_fonts}
\alias{load_all_artist_fonts}
\title{Load all Google Fonts used in artist styles}
\usage{
load_all_artist_fonts()
}
\description{
This function loads all necessary Google Fonts for the artist-inspired ggplot2 styles,
  preventing redundant loading when multiple style functions are called.
}
