% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_aux.R
\name{load_all_fonts}
\alias{load_all_fonts}
\title{Load Google Fonts used in artist styles}
\usage{
load_all_fonts(artist = "all")
}
\arguments{
\item{artist}{Character. One of the supported artists (e.g., "dali", "monet") or "all" to load all fonts. Default is "all".}
}
\description{
Load only the Google Fonts used in a specific artist's visual styles,
or all fonts if 'artist = "all"'. This avoids unnecessary font loading and speeds up rendering.
}
