% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_porcentaje.R
\name{plot_porcentaje}
\alias{plot_porcentaje}
\title{Gráfico de Barras de Porcentaje Redondeadas}
\usage{
plot_porcentaje(data, categoria, porcentaje, titulo = NULL)
}
\arguments{
\item{data}{Un data frame que contiene los datos a graficar.}

\item{categoria}{Un string que especifica el nombre de la columna que contiene las categorías (variable cualitativa) para el eje y.}

\item{porcentaje}{Un string que especifica el nombre de la columna que contiene los valores de porcentaje (variable numérica entre 0 y 1) para el eje x.}

\item{titulo}{Un string opcional que especifica el título principal del gráfico.}
}
\value{
Un objeto ggplot2 que representa el gráfico de barras de porcentaje redondeadas.
}
\description{
Genera un gráfico de barras horizontales con esquinas redondeadas, ideal para visualizar porcentajes o proporciones de forma clara y moderna.
}
