% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_styles.R
\name{preview_styles}
\alias{preview_styles}
\title{Previsualizar Estilos de Gráficos}
\usage{
preview_styles(style_name = "all", display_plots = TRUE)
}
\arguments{
\item{style_name}{Un string opcional para previsualizar un estilo específico. Si es "all" (por defecto), se previsualizan todos los estilos. Otras opciones válidas son: "da vinci"}

\item{display_plots}{Logical. Si es TRUE (por defecto), los gráficos se imprimen en el dispositivo actual uno por uno. Si es FALSE, se retorna una lista de objetos ggplot.}
}
\value{
Una lista de objetos ggplot si `display_plots` es FALSE. Si `display_plots` es TRUE, los gráficos se imprimen y la función retorna `NULL` de forma invisible.
}
\description{
Genera y muestra ejemplos de gráficos utilizando cada uno de los estilos definidos (Leonardo da Vinci, Artemisia Gentileschi, Cómic, Mapa Guapo, Porcentaje, Robótico, Elegante). Esto permite al usuario previsualizar y elegir el estilo que mejor se adapta a sus necesidades.
}
