% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_styles.R
\name{style_michelangelo}
\alias{style_michelangelo}
\title{Style for Michelangelo inspired plots}
\usage{
style_michelangelo(
  data,
  x,
  y,
  color_var = NULL,
  fill_var = NULL,
  label_var = NULL,
  title = "Magnificencia de Datos al Estilo de Miguel Ángel",
  subtitle = "Una composición de fuerza y detalle",
  caption = "Obra Maestra del Análisis",
  plot_type = c("scatter", "line", "column", "map"),
  work_inspired_by = c("david", "sistine_chapel", "pieta"),
  show_labels = FALSE,
  add_glow = FALSE,
  add_grid_lines = FALSE,
  show_background = TRUE,
  coord_flip = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{(Tidy-evaluated) Name of the column for the x-axis.}

\item{y}{(Tidy-evaluated) Name of the column for the y-axis.}

\item{color_var}{(Tidy-evaluated) Optional name of the column for color mapping. Defaults to NULL.}

\item{fill_var}{(Tidy-evaluated) Optional name of the column for fill mapping (for column charts). Defaults to NULL.}

\item{label_var}{(Tidy-evaluated) Optional name of the column for text labels (for maps or other plots). Defaults to NULL.}

\item{title}{The plot title.}

\item{subtitle}{The plot subtitle.}

\item{caption}{The plot caption.}

\item{plot_type}{The type of plot: "scatter", "line", or "column".}

\item{work_inspired_by}{Specific work for aesthetic inspiration: "david", "sistine_chapel", or "pieta".}

\item{show_labels}{Logical, whether to show data labels. Defaults to FALSE.}

\item{add_glow}{Logical, whether to add an outer glow effect. Defaults to FALSE.}

\item{add_grid_lines}{Logical, whether to show major grid lines. Defaults to FALSE.}

\item{show_background}{Logical, whether to show the panel background. Defaults to TRUE}

\item{coord_flip}{Logical, whether to flip coords in column chart (and so labels). Defaults to FALSE.}
}
\value{
A ggplot2 object.
}
\description{
Generates ggplot2 plots with aesthetic elements inspired by Michelangelo's works.
}
