% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_styles.R
\name{style_van_gogh}
\alias{style_van_gogh}
\title{Style for Vincent van Gogh inspired plots}
\usage{
style_van_gogh(
  data,
  x = NULL,
  y = NULL,
  color_var = NULL,
  fill_var = NULL,
  label_var = NULL,
  title = "Datos con la Pasión de Van Gogh",
  subtitle = "Un remolino de información",
  caption = "Impresiones de Datos",
  plot_type = c("column", "scatter", "line", "map"),
  work_inspired_by = c("starry_night", "sunflowers", "irises"),
  show_labels = FALSE,
  add_grid_lines = FALSE,
  show_background = TRUE,
  add_glow = FALSE,
  coord_flip = FALSE,
  text_size = 16,
  add_texture = NULL,
  canvas = NULL,
  add_margins = TRUE,
  add_filter = FALSE
)
}
\arguments{
\item{data}{A data frame or an sf object.}

\item{x}{A tidy-evaluated expression specifying the column name for the x-axis (for scatter, line, column plots).}

\item{y}{A tidy-evaluated expression specifying the column name for the y-axis (for scatter, line, column plots).}

\item{color_var}{An optional tidy-evaluated expression specifying the column name for color mapping. Defaults to NULL.}

\item{fill_var}{An optional tidy-evaluated expression specifying the column name for fill mapping (for column charts or map polygons). Defaults to NULL.}

\item{label_var}{An optional tidy-evaluated expression specifying the column name for text labels (for maps or other plots). Defaults to NULL.}

\item{title}{The plot title.}

\item{subtitle}{The plot subtitle.}

\item{caption}{The plot caption.}

\item{plot_type}{The type of plot: "scatter", "line", "column", or "map".}

\item{work_inspired_by}{Specific work for aesthetic inspiration: "starry_night", "sunflowers", or "irises".}

\item{show_labels}{Logical, whether to show data labels. Defaults to FALSE.}

\item{add_grid_lines}{Logical, whether to show major grid lines (only for non-map plots). Defaults to FALSE.}

\item{show_background}{Logical, whether to show the panel background (for non-map plots). Defaults to TRUE.}

\item{add_glow}{Logical, whether to add an outer glow effect. Defaults to FALSE.}

\item{coord_flip}{Logical, whether to flip coords in column chart (and so labels). Defaults to FALSE.}

\item{text_size}{Base size for text elements (titles, labels, legend). Labels and annotations scale proportionally from this value. Defaults to 16}

\item{add_texture}{Integer (1–3). Applies visual texture effects to geoms. Default: `NULL`.}

\item{canvas}{Integer (1–6). Adds canvas-style background image. Default: `NULL`.}

\item{add_margins}{Logical. Applies margins to plot and labs. Default: `TRUE`.}

\item{add_filter}{Experimental. Logical. Applies oil effect to the full graph. Default: `FALSE`.}
}
\value{
A ggplot2 object.
}
\description{
Generates ggplot2 plots with aesthetic elements inspired by Vincent van Gogh's works.
}
